stdlib/2and3/builtins: make dict.fromkeys(iterable) have None values, not Any
`dict.fromkeys()` has two forms:

1. `dict.fromkeys([1,2], "foo")` -> `{1: 'foo', 2: 'foo'}`
2. `dict.fromkeys([1,2])` -> `{1: None, 2: None}`

Previously, the second form gave the type `dict[int, None]`. Presumably
this is done to accommodate the common case where the values are filled
in later.

I think type safety should trump the minor inconvenience of needing to
use the longer form, so change it to None instead.