stubtest whitelists: clean up duplicate entries
This got duplicated a little during the merge. The source of the
confusion is that the issue was fixed in CPython and the changes were
backported to 3.7 and 3.8. However, the micro versions we currently use
on Linux don't have this fix.

Typeshed policy is to support the latest micro version, so I don't want
to hardcode micro versions into .travis.yml, since they'll never get
updated.