cachetools: fix caching decorators return type
CacheTools contains a collection of decorators (in cachetools.func)
which add a cache layer to any python function.

The decorated function has the same arguments and return types as the
original function.  The existing stub for these methods return 'None',
resulting in the following mypy errors:

```
error: "lru_cache" does not return a value
error: Untyped decorator makes function "_get_system_client" untyped
```

To correct this, I added a couple lines indicating that the output
was the same as the input.

mypy now passes without error.