socket: setblocking(int)
instead of `bool`, as there are many contradicting signatures:

<https://docs.python.org/3/library/socket.html#socket.socket.setblocking>
names the arguments `flags` and uses `false` in the descriptive text,
but uses `bool` in the given examples.

<https://github.com/python/cpython/blob/master/Modules/socketmodule.c#L149>
uses `bool` in the module documentation, but the internal wrapper in
<https://github.com/python/cpython/blob/master/Modules/socketmodule.c#L658>
uses `int` which is called by the public function in
<https://github.com/python/cpython/blob/master/Modules/socketmodule.c#L2783>
using `PyLong_AsLong()` to get the value of `flags` as an `int`.

So `int` is correct as `bool` is a subtype of it and gets promoted to it
automatically.

Signed-off-by: Philipp Hahn <hahn@univention.de>