Update __init__.pyi
If I read the source code correctly, the return type of `pkg_resources.load_entry_point` should be the same as the return type of `pkg_resources.EntryPoint.load`. The latter is annotated as `Any` but the former is annotated as `None` for some reason. This is bad because the return value is often expected to be as some sort of "main" function and gets invoked immediately -- but the `None` annotation will make type checkers emit the false alarm that `None` is not callable in those cases.