Stubs for stripe (#5770)
A library that provides convenient access to the Stripe API from applications. 
It includes a pre-defined set of classes for API resources that initialize themselves 
dynamically from API responses.

I started with stubgen, then cleaned up to address missing functions. Testing plan 
is to keep an eye on CI and fix any issues that arise.