Use TypeVar in add_constructor / add_multi_constructor
We need to use TypeVar in order to match the types in the
second parameter (construtor) and third parameter (Loader). Setting
`constructor: Callable[[BaseConstructor, None], Any]` and
`Loader: OptionalType[BaseConstructor]` doesn't work for when
`Loader` is a subclass `C` of `BaseConstructor`. This is because
`Callable[[C, None], Any]` cannot be substituted for
`Callable[[BaseConstructur], Any]`.