Change dependencies on types-cryptography, upstream includes type hints
Since version 3.4, the cryptography library itself ships with type
hints: https://cryptography.io/en/latest/changelog/#v3-4

This is the first step for removing types-cryptography.  Currently, if I
use `types-paramiko` and `cryptography`, it forces me to use
`types-cryptography` as well. It should be up to the end-user to decide
if they want upstream hints or from typeshed.

As discussed in the removal issue, I have added a dependency on
`types-paramiko` -> `cryptography` and `types-pyOpenSSL` ->
`cryptography`.