Remove covariance in mock.patch.multiple
The _T annotation for **kwargs would seem to imply that each of the kwarg values would be of the same type. However the interface conveys no such constraint: Any key can be mocked with any value, including DEFAULT.

A simple case that is improperly handled is `mock.patch.multiple(settings, A=1, B='foo')` since `1` and `'foo'` aren't consistent types.