Mark Cursor.rowcount as an `int`
DBAPI2 [specifies](https://www.python.org/dev/peps/pep-0249/#rowcount) that this is a "number" rather than specifically an int. It notes that a future version of the API might use `None` in place of `-1` to signal that there's no rowcount value to report. But that would presumably belong in a DBAPI version 3.