Make NoReturn a SpecialForm
This is consistent with the runtime definition: https://github.com/python/cpython/blob/main/Lib/typing.py#L434.

The previous definition was wrong; NoReturn is not and should not be equivalent to None. This fixes an issue in pyanalyze where it was interpreting NoReturn as equivalent to None.

cc @erictraut in case this affects pyright. mypy-primer will tell us what mypy thinks about it.