Revert "lru_cache preserves signature of wrapped function (#6221)" (#6356)
This reverts commit 8bda66a73725ff98919b8b57e600178591e948c2.

The change causes issues with ParamSpec implementations in type
checkers, at least pyright and my work-in-progress support for
ParamSpec in mypy. It's not yet clear how to fix the issues, so I
think that it's best to revert this, at least temporarily until we've
found a good solution. See #6347 for context.