Allow frame to be None in default_int_handler
Following the changes in https://github.com/python/typeshed/pull/5622
handlers are typed as taking an Optional frame but the
default_int_handler does not take None. This seems inconsistent
and pervents you from calling the default handler from a custom handler
without checking that the frame is not None first.