Hide non-runtime import
I recently ran into a user report that the type checker did not catch an invalid reference to `multiprocessing.SyncManager`, and this crashed at runtime because `SyncManager` is not available from the `multiprocessing` base module, but only from `multiprocessing.managers`. 

This leads me to suspect that the import in the stub file was only added to service type annotations in this module, and was not meant to reflect the availability of this name in the exports from this module. 

Proposing a patch to move this import to type checking only and rename it to avoid similar accidents in the future. I'm following the private type class naming convention already in the toplevel of this stub file, but don't have much context on these other classes, so leaving them untouched for now.