Add Y027 to list of ignored `.flake8` error codes
Refs https://github.com/PyCQA/flake8-pyi/pull/104

There are a number of stubs in typeshed that still support Python 2. Each subdirectory of stubs specifies whether or not it supports Python 2 in the `METADATA.toml` file. However, there's no easy way for flake8 to read those files, and I don't particularly like the idea of duplicating that information in the `.flake8` config file. As such, I think it's easiest if we just globally ignore this check for the whole of typeshed. We can still keep the equivalent check in `check_new_syntax.py`, where it's easier to dynamically specific filepaths that we want to check things for.