Remove unneeded `# noqa` comment
I added this in #6968 to make the CI go green for https://github.com/PyCQA/flake8-pyi/pull/109, but we've since decided that we only want to raise errors for `__str__` and `__repr__` if the signature is an exact match for `object.__str__`/`object.__repr__`. So this `# noqa` shouldn't be necessary.