Give precise return type for asyncio.gather(*aws)
Previously this used Any types. Caller `*args` looks like a fairly
common use case, so it would be nice to give a precise type in that
case. This could give a too general return type if `gather` is called
with a large number of positional arguments with non-uniform types,
but I'd expect this to be fairly rare.