Mark `Future._callbacks` as a `@property`
The _callbacks are actually initialized as an empty list and were typed correctly in typeshed
before python 3.7. After this it was not marked longer as a `@property`.

Although _callbacks is private and should not be used under normal circumstances, this popped
up in our unit tests where we validate if a function is actually set as a callback