Add pre-commit config for flake8
This makes it easier to lint your own work with `pre-commit`, but
doesn't change the CI config by adding this to the pre-commit.ci skip
list.

There's no way to automatically reflect the additional dependencies
from the requirements file, so the flake8 plugin list needs to be a
verbatim copy.

resolves #7991