Run stub_uploader tests in CI
stub_uploader is developing a bunch of validation, a lot of which must
live in that repo for security reasons. We don't want to find out about
this validation failing when stub uploading fails.

We also run it in the daily test since there are some checks that
(unfortunately, albeit unlikely) can be broken by new upstream releases.

This is blocked on https://github.com/typeshed-internal/stub_uploader/pull/60