feat: s2clientprotocol stub metadata
Constrain to version 5.* as that is the current of s2clientprotocol
which is being published.

The protobuf version is constrained to <4 for compatability reasons.
While it is unpinned in the requires for s2clientprotocol, all
downstream users actually need to end up pinning to <4 as the protoc
version (3.4.0) generates proto2 messages, which raise runtime errors
when used with protobuf>=4.