tkinter: Require passing in a callback function to `.after()`
Tkinter's `.after()` wraps Tcl's `after` command, which can be used in two ways: either as a blocking delay (like `time.sleep()` in Python), or to run a callback function in an event loop later (like `window.setTimeout` in javascript).

The blocking variant is useless in Python. It freezes your tkinter app and is equivalent to `time.sleep()` anyway.

I have seen other tkinter programmers using it twice, once by a beginner and once by an experienced tkinter user. In both cases, it was an accident and they forgot to pass in a callback function.